[modify] 사이드바 변수, 함수 분리
